+++ tests/files/repeat-block/output.lua
@@ -1,0 +1,3 @@
+repeat
+	x = x + 1
+until true