+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed a new line being added after the `until` token in a repeat block. The new line is now added at the end of the until expression.
 
 ## [0.1.0-alpha.2] - 2020-12-22
 ### Added